Допилено отображение фото друзей как ВК!

+++ vkClient2/Photo.swift
@@ -14,8 +14,8 @@ class Photo: Object {
     @objc  dynamic var id = 0
     @objc  dynamic var album_id = 0
     @objc  dynamic var owner_id = 0
-    @objc  dynamic var width = 0
-    @objc  dynamic var height = 0
+    @objc  dynamic var width = 0.0
+    @objc  dynamic var height = 0.0
     @objc  dynamic var date = 0
     @objc  dynamic var photo_75 = ""
     @objc  dynamic var photo_130 = ""
@@ -33,8 +33,8 @@ class Photo: Object {
         self.id = json["id"].intValue
         self.album_id = json["album_id"].intValue
         self.owner_id = json["owner_id"].intValue
-        self.width = json["width"].intValue
-        self.height = json["height"].intValue
+        self.width = json["width"].doubleValue
+        self.height = json["height"].doubleValue
         self.photo_75 = json["photo_75"].stringValue
         self.photo_2560 = json["photo_2560"].stringValue
         self.date = json["date"].intValue
